Add the time lapse options to the preferences dialogs. Change the sliders to QwtSlider in the tabs and the preferences. Add double and boolean preference elements and use it in the dialogs and the application.

+++ src/domain/domainfacade.h
@@ -398,13 +398,13 @@ public:
      * Get the live view fps of the project.
      * @return live view fps value.
      */
-    int getLiveViewFps();
+    double getLiveViewFps();
 
     /**
      * Set the live view fps of the project.
      * @param newValue The new live view fps value.
      */
-    void setLiveViewFps(int newValue);
+    void setLiveViewFps(double newValue);
 
     /**************************************************************************
      * Video export preferences
